requirements.txt
ollama
psycopg2-binary
python-dotenv